+++ gcpy/cstools.py
@@ -720,9 +720,7 @@ def is_cubed_sphere_diag_grid(data):
     True if the grid has History diagnostic dimensions,
     False otherwise.
     """
-    if "nf" in data.dims:
-        return True
-    return False
+    return "nf" in data.dims
 
 
 def is_cubed_sphere_rst_grid(data):
@@ -746,12 +744,17 @@ def is_cubed_sphere_rst_grid(data):
     # internal state variables in GCHP. A more robust back-up check
     # could be to see if all the lats and lons are integer, since
     # that will be the case with the GCHP restart file format.
-    if "lat" in data.dims:
-        if data.dims["lat"] == data.dims["lon"] * 6:
-            return True
-    if "SPC_" in data.data_vars.keys():
-        return True
-    return False
+
+    # NOTE: in DataArray objects, dims is a tuple but not a dict!
+    # Comparing the len of the lat & lon coords will work for both.
+    if "lat" in data.coords:
+         return len(data.coords["lat"]) == len(data.coords["lon"]) * 6
+
+    # Dataset: search data.data_vars for "SPC_"
+    # DataArray: search data.name for "SPC_"
+    if isinstance(data, xr.Dataset):
+        return "SPC_" in data.data_vars.keys()
+    return "SPC_" in data.name
 
 
 def get_cubed_sphere_res(data):
@@ -776,9 +779,13 @@ def get_cubed_sphere_res(data):
 
     if not is_cubed_sphere(data):
         return 0
+
+    # NOTE: In Dataset objects "dims" is a dict, but in DataArray
+    # objects "dims" is a tuple.  Returning the length of the
+    # corresponding coords array should work in both cases.
     if is_cubed_sphere_rst_grid(data):
-        return data.dims["lon"]
-    return data.dims["Xdim"]
+        return len(data.coords["lon"])
+    return len(data.coords["Xdim"])
 
 
 def is_gchp_lev_positive_down(data):